added versioning to crypto package

+++ crypto/src/main/java/de/tuberlin/tfdacmacs/crypto/pairing/data/keys/UserAttributeValueUpdateKey.java
@@ -4,7 +4,7 @@ import it.unisa.dia.gas.jpbc.Element;
 import lombok.NonNull;
 
 public class UserAttributeValueUpdateKey extends UserUpdateKey {
-    public UserAttributeValueUpdateKey(@NonNull String userId, @NonNull Element updateKey, int targetVersion) {
+    public UserAttributeValueUpdateKey(@NonNull String userId, @NonNull Element updateKey, long targetVersion) {
         super(userId, updateKey, targetVersion);
     }
 }
